app/requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.31
qrcode[pil]
pytz==2022.1
